scripts/python/requirements.txt
supabase
ollama
python-dotenv
